Add Config struct for env loading
